Refactor for documentable filters.
